+++ example/android/app/src/main/kotlin/dev/xx/video_view_example/MainActivity.kt
@@ -2,4 +2,4 @@ package dev.xx.video_view_example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity()
+class MainActivity : FlutterActivity()